Fix the TypeError in Maya 2019, CentOS
# TypeError: # 'PySide2.QtGui.QPixmap.grabWindow' called with wrong argument types:
#   PySide2.QtGui.QPixmap.grabWindow(int, int, int, int, int)
# Supported signatures:
  PySide2.QtGui.QPixmap.grabWindow(quintptr, int = 0, int = 0, int = -1, int = -1)

+++ python/screen_grab/screen_grab.py
@@ -322,7 +322,7 @@ def get_desktop_pixmap(rect):
     :rtype: :class:`~PySide.QtGui.QPixmap`
     """
     desktop = QtGui.QApplication.desktop()
-    return QtGui.QPixmap.grabWindow(desktop.winId(), rect.x(), rect.y(),
+    return QtGui.QPixmap.grabWindow(long(desktop.winId()), rect.x(), rect.y(),
                                     rect.width(), rect.height())
